requirements.txt
tensorflow
gradio
Pillow
numpy
